+++ refinedstorage2-network/src/test/java/com/refinedmods/refinedstorage2/api/network/impl/node/exporter/RoundRobinExporterNetworkNodeTest.java
@@ -29,6 +29,7 @@ import static org.mockito.Mockito.verify;
 class RoundRobinExporterNetworkNodeTest extends AbstractExporterNetworkNodeTest {
     private Runnable listener;
 
+    @Override
     @BeforeEach
     void setUp() {
         listener = mock(Runnable.class);